test: move figure data JS back into test root and ignore in config
