+++ packages/components/bolt-figure/__tests__/figure-data.js
